Separate instructor card components
